requirements.txt
fastapi
uvicorn[standard]
instaloader
groq
python-dotenv
jinja2
python-multipart
httpx
